refactors function to generatesPrompts.js file, adds style.css, new team.html output
