+++ index.js
@@ -1,18 +1,19 @@
-const Manager = require("./lib/Manager");
-const Engineer = require("./lib/Engineer");
-const Intern = require("./lib/Intern");
-const inquirer = require("inquirer");
-const path = require("path");
-const fs = require("fs");
-const PromptBuilder = require("./lib/PromptBuilder");
 
-const OUTPUT_DIR = path.resolve(__dirname, "output");
-const outputPath = path.join(OUTPUT_DIR, "team.html");
 
-const render = require("./src/page-template.js");
+const generatesPrompts = require("./lib/generatesPrompts");
 
+function init() {
+
+    console.log("Answer the following questions to generate your Team Profile file");
+
+    generatesPrompts();
+
+}
+
+init();
 
 // TODO: Write Code to gather information about the development team members, and render the HTML file.
+
 
 //console log welcome message
 //create teamdata array
@@ -41,96 +42,5 @@ const render = require("./src/page-template.js");
 // THEN BRANCH into either |||create eng/int -> push to arr -> RETURN to options prompt (recursive? https://github.com/SBoudrias/Inquirer.js/blob/master/packages/inquirer/examples/recursive.js)||| OR |||html generation -> file write -> exit|||
 // https://github.com/SBoudrias/Inquirer.js/blob/master/packages/inquirer/examples/hierarchical.js
 
-function init() {
-    console.log("Answer the following questions to generate your Team Profile file");
-
-    const teamDataArr = [];
-
-    const managerPromptArr =
-        new PromptBuilder("team manager")
-            .createNamePrompt()
-            .createIdPrompt().
-            createEmailPrompt()
-            .createOfficeNumberPrompt()
-            .createPromptsArr();
-
-    inquirer.prompt(managerPromptArr).then(data => {
-        // console.log(data);
-        const { name, id, email, officeNumber } = data;
-        const managerObj = new Manager(name, id, email, officeNumber);
-        teamDataArr.push(managerObj);
-        // console.log(teamDataArr);
-        generatesMenuPrompt();
-    });
-
-    function generatesMenuPrompt() {
-
-        const optionsPromptArr =
-        new PromptBuilder()
-            .createOptionsPrompt()
-            .createPromptsArr();
-
-        inquirer.prompt(optionsPromptArr)
-        .then( data => {
-            switch (data.options) {
-                case "Engineer":
-                    generatesEngineerPrompt();
-                    break
-                    case "Intern":
-                        generatesInternPrompt();
-                        break
-                        case "Finish building the team - Generate Team Profile":
-                            console.log(teamDataArr);
-                            const renderedHTML = render(teamDataArr);
-                            if (!fs.existsSync(OUTPUT_DIR)) {
-                                fs.mkdir(OUTPUT_DIR, err => err ? console.error(err) : "");
-                            }
-                            fs.writeFile(outputPath, renderedHTML, err => err ? console.error(err) : "");
-                            process.exit;
-                            // startsExitProcess();
-            }
-        })
-    }
-
-    function generatesEngineerPrompt() {
-
-        const engineerPromptArr =
-        new PromptBuilder("engineer")
-            .createNamePrompt()
-            .createIdPrompt().
-            createEmailPrompt()
-            .createGithubPrompt()
-            .createPromptsArr();
-
-        inquirer.prompt(engineerPromptArr)
-        .then( data => {
-            const {name, id, email, github} = data;
-            const engineerObj = new Engineer(name, id, email, github);
-            teamDataArr.push(engineerObj);
-            generatesMenuPrompt();
-        })
-    }
-
-    function generatesInternPrompt() {
-
-        const internPromptArr =
-        new PromptBuilder("intern")
-            .createNamePrompt()
-            .createIdPrompt().
-            createEmailPrompt()
-            .createSchoolPrompt()
-            .createPromptsArr();
-
-        inquirer.prompt(internPromptArr)
-        .then( data => {
-            const {name, id, email, school} = data;
-            const internObj = new Intern(name, id, email, school);
-            teamDataArr.push(internObj);
-            generatesMenuPrompt();
-        })
-    }
-}
-
-init();
 
 // console.log(OUTPUT_DIR);
