+++ output/style.css
@@ -1,0 +1,33 @@
+.team-heading {
+    background-image: linear-gradient(45deg, blue, grey);
+    color: aliceblue;
+    border-radius: 0em;
+}
+
+.card-body {
+    display: flex;
+    flex-direction: column;
+    justify-content: space-around;
+    background-color: lightskyblue;
+}
+
+.team-area {
+    display: flex;
+    justify-content: center;
+    flex-direction: row;
+    flex-wrap: wrap;
+    margin: 10% auto;
+}
+
+.employee-card {
+    margin: 1rem;
+    border-radius: 0.6rem;
+}
+
+.list-group-item {
+    background-color: lightcyan;
+}
+
+body {
+    background-color: lightsteelblue;
+}
