+++ output/team.html
@@ -32,36 +32,50 @@
         </div>
         <div class="card-body">
             <ul class="list-group">
-                <li class="list-group-item">ID: 324</li>
-                <li class="list-group-item">Email: <a href="mailto:bob@company.com">bob@company.com</a></li>
-                <li class="list-group-item">Office number: 324</li>
+                <li class="list-group-item">ID: 12</li>
+                <li class="list-group-item">Email: <a href="mailto:bob@c.com">bob@c.com</a></li>
+                <li class="list-group-item">Office number: 4432</li>
             </ul>
         </div>
     </div>
         
         <div class="card employee-card">
     <div class="card-header">
-        <h2 class="card-title">bobby</h2>
+        <h2 class="card-title">bonny</h2>
         <h3 class="card-title"><i class="fas fa-glasses mr-2"></i>Engineer</h3>
     </div>
     <div class="card-body">
         <ul class="list-group">
-            <li class="list-group-item">ID: 1324</li>
-            <li class="list-group-item">Email: <a href="mailto:bobby@company.com">bobby@company.com</a></li>
-            <li class="list-group-item">GitHub: <a href="https://github.com/bobby1" target="_blank" rel="noopener noreferrer">bobby1</a></li>
+            <li class="list-group-item">ID: 934</li>
+            <li class="list-group-item">Email: <a href="mailto:bonny@c.com">bonny@c.com</a></li>
+            <li class="list-group-item">GitHub: <a href="https://github.com/bonnyM" target="_blank" rel="noopener noreferrer">bonnyM</a></li>
         </ul>
     </div>
 </div>
         
         <div class="card employee-card">
     <div class="card-header">
-        <h2 class="card-title">bonny</h2>
+        <h2 class="card-title">bobby</h2>
         <h3 class="card-title"><i class="fas fa-user-graduate mr-2"></i>Intern</h3>
     </div>
     <div class="card-body">
         <ul class="list-group">
-            <li class="list-group-item">ID: 324</li>
-            <li class="list-group-item">Email: <a href="mailto:bonny@company.com">bonny@company.com</a></li>
+            <li class="list-group-item">ID: 68</li>
+            <li class="list-group-item">Email: <a href="mailto:bobby@c.com">bobby@c.com</a></li>
+            <li class="list-group-item">School: www school</li>
+        </ul>
+    </div>
+</div>
+        
+        <div class="card employee-card">
+    <div class="card-header">
+        <h2 class="card-title">jane</h2>
+        <h3 class="card-title"><i class="fas fa-user-graduate mr-2"></i>Intern</h3>
+    </div>
+    <div class="card-body">
+        <ul class="list-group">
+            <li class="list-group-item">ID: 934</li>
+            <li class="list-group-item">Email: <a href="mailto:jane@c.com">jane@c.com</a></li>
             <li class="list-group-item">School: w3schools</li>
         </ul>
     </div>
@@ -69,14 +83,14 @@
         
         <div class="card employee-card">
     <div class="card-header">
-        <h2 class="card-title">john</h2>
+        <h2 class="card-title">tieno</h2>
         <h3 class="card-title"><i class="fas fa-user-graduate mr-2"></i>Intern</h3>
     </div>
     <div class="card-body">
         <ul class="list-group">
-            <li class="list-group-item">ID: 4325</li>
-            <li class="list-group-item">Email: <a href="mailto:john@company.com">john@company.com</a></li>
-            <li class="list-group-item">School: school of learning</li>
+            <li class="list-group-item">ID: 94</li>
+            <li class="list-group-item">Email: <a href="mailto:tieno@gmail.com">tieno@gmail.com</a></li>
+            <li class="list-group-item">School: CS school</li>
         </ul>
     </div>
 </div>
